+++ Runtime/Core/Predictor.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Cinemachine.Utility
@@ -5,6 +6,11 @@ namespace Cinemachine.Utility
     public class PositionPredictor
     {
         Vector3 m_Position;
+
+        GaussianWindow1D_Vector3 m_Velocity = new GaussianWindow1D_Vector3(kSmoothingDefault);
+        GaussianWindow1D_Vector3 m_Accel = new GaussianWindow1D_Vector3(kSmoothingDefault);
+
+        float mLastVelAddedTime = 0;
 
         const float kSmoothingDefault = 10;
         float mSmoothing = kSmoothingDefault;
@@ -22,10 +28,6 @@ namespace Cinemachine.Utility
                 }
             }
         }
-
-        GaussianWindow1D_Vector3 m_Velocity = new GaussianWindow1D_Vector3(kSmoothingDefault);
-        GaussianWindow1D_Vector3 m_Accel = new GaussianWindow1D_Vector3(kSmoothingDefault);
-
         public bool IsEmpty { get { return m_Velocity.IsEmpty(); } }
 
         public void ApplyTransformDelta(Vector3 positionDelta)
@@ -39,38 +41,44 @@ namespace Cinemachine.Utility
             m_Accel.Reset();
         }
 
-        public void AddPosition(Vector3 pos)
+        public void AddPosition(Vector3 pos, float deltaTime, float lookaheadTime)
         {
-            if (IsEmpty)
+            if (deltaTime < UnityVectorExtensions.Epsilon)
+                Reset();
+            else if (IsEmpty)
                 m_Velocity.AddValue(Vector3.zero);
-            else if (Time.deltaTime > Vector3.kEpsilon)
+            else
             {
-                Vector3 vel = m_Velocity.Value();
-                Vector3 vel2 = (pos - m_Position) / Time.deltaTime;
-                m_Velocity.AddValue(vel2);
-                m_Accel.AddValue(vel2 - vel);
+                Vector3 vel = (pos - m_Position) / deltaTime;
+                if (vel.sqrMagnitude > UnityVectorExtensions.Epsilon)
+                {
+                    Vector3 vel0 = m_Velocity.Value();
+                    float now = Time.time;
+                    if (vel.sqrMagnitude >= vel0.sqrMagnitude
+                        || Vector3.Angle(vel, vel0) > 10
+                        || now > mLastVelAddedTime + lookaheadTime)
+                    {
+                        m_Velocity.AddValue(vel);
+                        m_Accel.AddValue(vel - vel0);
+                        mLastVelAddedTime = now;
+                    }
+                }
             }
             m_Position = pos;
         }
 
+        public Vector3 PredictPositionDelta(float lookaheadTime)
+        {
+            Vector3 vel = m_Velocity.IsEmpty() ? Vector3.zero : m_Velocity.Value();
+            Vector3 accel = m_Accel.IsEmpty() ? Vector3.zero : m_Accel.Value();
+            Vector3 delta = vel * lookaheadTime;
+            delta += accel * lookaheadTime * lookaheadTime * 0.5f;
+            return delta;
+        }
+
         public Vector3 PredictPosition(float lookaheadTime)
         {
-            Vector3 pos = m_Position;
-            if (Time.deltaTime > Vector3.kEpsilon)
-            {
-                int numSteps = Mathf.Min(Mathf.RoundToInt(lookaheadTime / Time.deltaTime), 6);
-                float dt = lookaheadTime / numSteps;
-                Vector3 vel = m_Velocity.IsEmpty() ? Vector3.zero : m_Velocity.Value();
-                Vector3 accel = m_Accel.IsEmpty() ? Vector3.zero : m_Accel.Value();
-                for (int i = 0; i < numSteps; ++i)
-                {
-                    pos += vel * dt;
-                    Vector3 vel2 = vel + (accel * dt);
-                    accel = Quaternion.FromToRotation(vel, vel2) * accel;
-                    vel = vel2;
-                }
-            }
-            return pos;
+            return m_Position + PredictPositionDelta(lookaheadTime);
         }
     }