+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: collider minimum occlusion time and smoothing time interaction
 - Bugfix: TargetGroup.RemoveMember error (fogbugz 1119028)
 - Bugfix: TargetGroup member lerping jerk when member weight near 0
+- Bugfix: Transposer angular damping should be 0 only if binding mode not LockToTarget
 
 ## [2.3.3] - 2019-01-08
 ### Temporary patch to get around a Unity bug in conditional dependencies