+++ CHANGELOG.md
@@ -4,9 +4,8 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.3.4-preview.3] - 2019-02-18
+## [2.3.4-preview.4] - 2019-04-18
 ### PostProcessingV3 and bugfixes
-- Added conditional compilation to support removal of optional physics and physics2D modules
 - Added CinemachineCore.GetBlendOverride delegate to allow applications to override any vcam blend when it happens
 - When a blend is cancelled by the opposite blend, reduce the blend time
 - Added support for PostProcessing V3 - now called CinemachineVolumeSttings
@@ -14,8 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Timeline won't auto-create CM brains when something dragged onto it
 - Confiner: Improvement in automatic path invalidation when number of path points path changes
 - Added CinemachineInpuitAxisDriver utility for overriding the default AxisState behaviour
-- CinemachineCameraOffset: added customazable stage for when to apply the offset
+- CinemachineCameraOffset: added customizable stage for when to apply the offset
 - Added Loop option to BlendList Camera
+- Improved Lookahed: does not automatically recenter
+- Bugfix: Collider must consider actual target position, not lookahead position
 - Bugfix: FreeLook heading RecenterNow was not working
 - Bugfix: lookahead now takes the overridden Up into account
 - Bugfix: screen composer guides drawn in wrong place for picture-in-picture